+++ vehicle/src/Vehicle/Compile/Type/Meta/Substitution.hs
@@ -119,21 +119,11 @@ instance MetaSubstitutable m builtin (UnificationConstraint builtin) where
 
 instance MetaSubstitutable m builtin (InstanceConstraint builtin) where
   subst s (Resolve origin m r g) =
-    Resolve <$> subst s origin <*> substMetaID s m <*> pure r <*> subst s g
+    Resolve <$> subst s origin <*> subst s m <*> pure r <*> subst s g
 
 instance MetaSubstitutable m builtin (InstanceGoal builtin) where
   subst s (InstanceGoal t h spine) =
     InstanceGoal t h <$> subst s spine
-
--- This is a massive hack, and only works because we only have instance resolution
--- for types in the loss typing subsystem which doesn't use dependently types.
-substMetaID :: (MonadLogger m) => MetaSubstitution builtin -> MetaID -> m MetaID
-substMetaID metaSubst m =
-  case MetaMap.lookup m metaSubst of
-    Nothing -> return m
-    Just other -> case normalised other of
-      VMeta m2 [] -> substMetaID metaSubst m2
-      _ -> return m
 
 instance MetaSubstitutable m builtin (Constraint builtin) where
   subst s = \case